requirements.txt
ariadne==0.21
asyncpg==0.29.0
black==23.10.1
fastapi==0.108.0
flake8==6.1.0
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-jose==3.3.0
tortoise-orm==0.20.0
uvicorn==0.25.0

